Minor tweaks for performance

+++ DebyeWolfIntegral.py
@@ -49,6 +49,8 @@ def TightFocus(InputField_x,InputField_y,dx,wavelength,n_homogenous,FocusDepth,M
 
     For each x,y, calculate three integrals for Ex, Ey, Ez
     '''
+    InputField_x = InputField_x.astype('complex64')
+    InputField_y = InputField_y.astype('complex64')
 
     MeasurementPlane_z = -MeasurementPlane_z    # Origin is at focus and z axis is along the direction of propagation. See Fig 1, [1]
     xy_cells = np.shape(InputField_x)[0]
@@ -79,7 +81,7 @@ def TightFocus(InputField_x,InputField_y,dx,wavelength,n_homogenous,FocusDepth,M
     f = 1/(dx*xy_cells)*indices
     k = 2*np.pi*n_homogenous/wavelength
     k0 = 2*np.pi/wavelength
-    H = np.exp(1j*FocusDepth*np.emath.sqrt((k)**2-kxkx**2-kyky**2))
+    H = np.exp(1j*FocusDepth*np.emath.sqrt((k)**2-kxkx**2-kyky**2)).astype('complex64')
     InputField_x_propagated = iFFT2(FFT2(InputField_x)*H)
     InputField_y_propagated = iFFT2(FFT2(InputField_y)*H)
 
@@ -159,15 +161,15 @@ def TightFocus(InputField_x,InputField_y,dx,wavelength,n_homogenous,FocusDepth,M
     xx_final, yy_final = np.meshgrid(target_dx*indices,target_dx*indices,indexing='ij')
     mask = (np.abs(theta)<np.pi/2-0.01).astype('int')   # Everything beyond the theta = pi/2 region shoul be zero.
 
-    Ex = RegularGridInterpolator((dx_step1*index_step1,dx_step1*index_step1),prefactor*FFT2(np.pad(mask*np.exp(1j*kz*MeasurementPlane_z)*Ax/kz,pad)), bounds_error = False, fill_value = 0, method='linear')((xx_final, yy_final))
-    Ey = RegularGridInterpolator((dx_step1*index_step1,dx_step1*index_step1),prefactor*FFT2(np.pad(mask*np.exp(1j*kz*MeasurementPlane_z)*Ay/kz,pad)), bounds_error = False, fill_value = 0, method='linear')((xx_final, yy_final))
-    Ez = RegularGridInterpolator((dx_step1*index_step1,dx_step1*index_step1),prefactor*FFT2(np.pad(mask*np.exp(1j*kz*MeasurementPlane_z)*Az/kz,pad)), bounds_error = False, fill_value = 0, method='linear')((xx_final, yy_final))
+    Ex = RegularGridInterpolator((dx_step1*index_step1,dx_step1*index_step1),prefactor*FFT2(np.pad((mask*np.exp(1j*kz*MeasurementPlane_z)*Ax/kz).astype('complex64'),pad)), bounds_error = False, fill_value = 0, method='linear')((xx_final, yy_final))
+    Ey = RegularGridInterpolator((dx_step1*index_step1,dx_step1*index_step1),prefactor*FFT2(np.pad((mask*np.exp(1j*kz*MeasurementPlane_z)*Ay/kz).astype('complex64'),pad)), bounds_error = False, fill_value = 0, method='linear')((xx_final, yy_final))
+    Ez = RegularGridInterpolator((dx_step1*index_step1,dx_step1*index_step1),prefactor*FFT2(np.pad((mask*np.exp(1j*kz*MeasurementPlane_z)*Az/kz).astype('complex64'),pad)), bounds_error = False, fill_value = 0, method='linear')((xx_final, yy_final))
 
     # Normalization
     n0 = np.sum(np.abs(Ex)**2+np.abs(Ey)**2+np.abs(Ez)**2)*out_dx**2
-    Ex = Ex/n0
-    Ey = Ey/n0
-    Ez = Ez/n0
+    Ex = (Ex/n0).astype('complex64')
+    Ey = (Ey/n0).astype('complex64')
+    Ez = (Ez/n0).astype('complex64')
 
     if ScalingFactor2 <0.1:
         wrn = 'High output interpolation factor ('+str(1/ScalingFactor2)+'): Increase xy_cells or dx'
